Find a decennial or ACS census table with variables of interest

This function uses fuzzy matching to help identify tables from the census which contain variables of interest. Matched table codes are printed out, along with the Census-provided table description, the parsed variable names, and example table cells. The website https://censusreporter.org/ may also be useful in finding variables.

Usage

find_dec_table(..., show = 2)

find_acs_table(..., show = 4)

Arguments

...

Variables to look for. These can be length-1 character vectors, or, for convenience, can be left unquoted (see examples).

show

How many matching tables to show. Increase this to show more possible matches, at the cost of more output. Negative values will be converted to positive but will suppress any printing.

Value

The codes for the top show tables, invisibly.

Examples

find_dec_table("sex", "age")
#> 
#> ── Top 2 matching tables ───────────────────────────────────────────────────────
#> 
#>  P012  - SEX BY AGE
#> Parsed variables:
#> • sex
#> • age
#> • race_ethnicity
#> Example values:
#> • male / 67 to 69 years / two or more races
#> • female / total / native hawaiian and other pacific islander alone
#> • female / 25 to 29 years / total
#> 
#>  PCT012  - SEX BY AGE
#> Parsed variables:
#> • sex
#> • age
#> • race_ethnicity
#> Example values:
#> • female / 70 years / asian alone, not hispanic or latino
#> • male / 28 years / two or more races
#> • female / 110 years and over / black or african american alone
find_dec_table(tenure, race)
#> 
#> ── Top 2 matching tables ───────────────────────────────────────────────────────
#> 
#>  H014  - TENURE BY RACE OF HOUSEHOLDER
#> Parsed variables:
#> • tenure
#> • race_of_householder
#> Example values:
#> • owner occupied / householder who is asian alone
#> • owner occupied / householder who is white alone
#> • renter occupied / householder who is two or more races
#> 
#>  H016  - TENURE BY HOUSEHOLD SIZE
#> Parsed variables:
#> • tenure
#> • household_size
#> • race_ethnicity
#> Example values:
#> • owner occupied / total / black or african american alone householder
#> • renter occupied / 6-person household / some other race alone householder
#> • owner occupied / 7-or-more-person household / white alone householder
find_acs_table("income", "sex", show=10)
#> 
#> ── Top 10 matching tables ──────────────────────────────────────────────────────
#> 
#>  B17008  - AGGREGATE INCOME DEFICIT (DOLLARS) IN THE PAST 12 MONTHS OF
#> UNRELATED INDIVIDUALS BY SEX
#> Parsed variables:
#> • sex
#> Example values:
#> • male (dollars)
#> • total
#> • female (dollars)
#> 
#>  B19216  - AGGREGATE NONFAMILY HOUSEHOLD INCOME IN THE PAST 12 MONTHS (IN 2019
#> INFLATION-ADJUSTED DOLLARS) BY SEX OF HOUSEHOLDER BY LIVING ALONE BY AGE OF
#> HOUSEHOLDER
#> Parsed variables:
#> • sex_of_householder
#> • living_alone
#> • age_of_householder
#> Example values:
#> • male householder (dollars) / not living alone (dollars) / householder 65
#> years and over (dollars)
#> • male householder (dollars) / not living alone (dollars) / householder 15 to
#> 64 years (dollars)
#> • total / total / total
#> 
#>  B21004  - MEDIAN INCOME IN THE PAST 12 MONTHS (IN 2019 INFLATION-ADJUSTED
#> DOLLARS) BY VETERAN STATUS BY SEX FOR THE CIVILIAN POPULATION 18 YEARS AND OVER
#> WITH INCOME
#> Parsed variables:
#> • veteran_status
#> • sex_for_the_civilian_population_18_years_over_with_income
#> Example values:
#> • veteran / female
#> • nonveteran / female
#> • veteran / male
#> 
#>  B19326  - MEDIAN INCOME IN THE PAST 12 MONTHS (IN 2019 INFLATION-ADJUSTED
#> DOLLARS) BY SEX BY WORK EXPERIENCE IN THE PAST 12 MONTHS FOR THE POPULATION 15
#> YEARS AND OVER WITH INCOME
#> Parsed variables:
#> • median_income_in_the_past_12_months
#> • sex
#> •
#> work_experience_in_the_past_12_months_for_the_population_15_years_over_with_income
#> Example values:
#> • female / total (dollars) / worked full-time, year-round in the past 12 months
#> (dollars)
#> • male / total (dollars) / other (dollars)
#> • male / total (dollars) / total
#> 
#>  B19083  - GINI INDEX OF INCOME INEQUALITY
#> Parsed variables:
#> • gini_index_of_income_inequality
#> Example values:
#> • gini index
#> 
#>  B19325  - SEX BY WORK EXPERIENCE IN THE PAST 12 MONTHS BY INCOME IN THE PAST
#> 12 MONTHS (IN 2019 INFLATION-ADJUSTED DOLLARS) FOR THE POPULATION 15 YEARS AND
#> OVER
#> Parsed variables:
#> • sex
#> • work_experience_in_the_past_12_months
#> • income_in_the_past_12_months_for_the_population_15_years_over
#> •
#> sex.work_experience_in_the_past_12_months.income_in_the_past_12_months_for_the_population_15_years_over_sub
#> Example values:
#> • female / worked full-time, year-round in the past 12 months / with income /
#> $55,000 to $64,999
#> • male / other / with income / total
#> • female / other / with income / $7,500 to $9,999
#> 
#>  B19215  - MEDIAN NONFAMILY HOUSEHOLD INCOME IN THE PAST 12 MONTHS (IN 2019
#> INFLATION-ADJUSTED DOLLARS) BY SEX OF HOUSEHOLDER BY LIVING ALONE BY AGE OF
#> HOUSEHOLDER
#> Parsed variables:
#> • median_nonfamily_household_income_in_the_past_12_months
#> • sex_of_householder
#> • living_alone
#> • age_of_householder
#> •
#> median_nonfamily_household_income_in_the_past_12_months.sex_of_householder.living_alone.age_of_householder_sub
#> Example values:
#> • male householder / total (dollars) / total / total / total
#> • male householder / total (dollars) / not living alone / total (dollars) /
#> householder 15 to 64 years (dollars)
#> • male householder / total (dollars) / living alone / total (dollars) /
#> householder 15 to 64 years (dollars)
#> 
#>  B19082  - SHARES OF AGGREGATE HOUSEHOLD INCOME BY QUINTILE
#> Parsed variables:
#> • shares_of_aggregate_household_income
#> • quintile
#> Example values:
#> • top 5 percent / total
#> • quintile share of aggregate income / third quintile
#> • quintile share of aggregate income / lowest quintile
#> 
#>  B25119  - MEDIAN HOUSEHOLD INCOME THE PAST 12 MONTHS (IN 2019
#> INFLATION-ADJUSTED DOLLARS) BY TENURE
#> Parsed variables:
#> • tenure
#> Example values:
#> • renter occupied (dollars)
#> • total
#> • owner occupied (dollars)
#> 
#>  B19013  - MEDIAN HOUSEHOLD INCOME IN THE PAST 12 MONTHS (IN 2019
#> INFLATION-ADJUSTED DOLLARS)
#> Parsed variables:
#> • race_ethnicity
#> Example values:
#> • white alone, not hispanic or latino householder
#> • american indian and alaska native alone householder
#> • some other race alone householder
find_acs_table("heath care", show=-1)