Download data from a decennial census or ACS table

Download data from a decennial census or ACS table

Usage

get_dec_table(geography, table, ..., drop_total = FALSE)

get_acs_table(
  geography,
  table,
  survey = c("acs5", "acs3", "acs1"),
  ...,
  drop_total = FALSE
)

Arguments

geography

The geography level to download data for. Usually one of state, county, tract, block group, block, zcta, etc. Consult https://walker-data.com/tidycensus/articles/basic-usage.html#geography-in-tidycensus for more information.

table

The table code to download. See find_dec_table() for help identifying a table of interest. Note: some tables are split into A/B/C/etc. versions by race; this function unifies all of these tables under one code. So, for example, use P012, not P012A.

...

Arguments passed to tidycensus::get_decennial() or tidycensus::get_acs(), e.g. year, state, county, geometry.

drop_total

Whether to filter out variables which are totals across another variable. Recommended only after inspection of the underlying table.

survey

For ACS data, whether to use the one-year, three-year, or five-year survey (the default). Note that some tables from find_acs_table() may only be available for acs5.

Value

A tibble of census data in tidy format, with columns GEOID, NAME, variable (containing the Census variable code), value or estiamte, moe in the case of ACS tables, and additional factor columns specific to the table.

Examples

if (FALSE) {
get_dec_table("state", "P003")
get_dec_table("state", "H002")
get_dec_table("county", "H002", state="WA", drop_total=TRUE)

get_acs_table("county subdivision", "B09001", state="WA", county="King")
}