Find a decennial or ACS census table with variables of interest

This function uses fuzzy matching to help identify tables from the census which contain variables of interest. Matched table codes are printed out, along with the Census-provided table description, the parsed variable names, and example table cells. The website https://censusreporter.org/ may also be useful in finding variables.

Usage

cens_find_dec(..., show = 2)

cens_find_acs(..., show = 4)

Arguments

...

Variables to look for. These can be length-1 character vectors, or, for convenience, can be left unquoted (see examples).

show

How many matching tables to show. Increase this to show more possible matches, at the cost of more output. Negative values will be converted to positive but will suppress any printing.

Value

The codes for the top show tables, invisibly.

Examples

cens_find_dec("sex", "age")
#> Error in match_tables(variables, tables_sf1): object 'tables_sf1' not found
cens_find_dec(tenure, race)
#> Error in match_tables(variables, tables_sf1): object 'tables_sf1' not found
cens_find_acs("income", "sex", show=3)
#> Error in match_tables(variables, tables_acs, complex_pen = 0.1): object 'tables_acs' not found
cens_find_acs("heath care", show=-1)
#> Error in match_tables(variables, tables_acs, complex_pen = 0.1): object 'tables_acs' not found